// payload-types.ts
export interface User {
  id: string;
  name?: string | null;
  username?: string | null;
  bio?: string | null;
  avatar?: string | Media | null;
  authProvider: "local" | "google" | "facebook" | "x";
  sub?: string | null;
  roles?: ("admin" | "user")[] | null;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  password: string | null;
}
// ...
export interface Workflow {
  id: string;
  name: string;
  serviceName: string;
  directions: string;
  cost: number;
  desc?: string | null;
  cover?: string | Media | null;
  data:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  updatedAt: string;
  createdAt: string;
  _status?: ("draft" | "published") | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string;
  alt: string;
  caption?:
    | {
        [k: string]: unknown;
      }[]
    | null;
  prefix?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "pointTransferLogs".
 */
export interface PointTransferLog {
  id: string;
  user: string | User;
  workflow?: (string | null) | Workflow;
  event: string;
  status: string;
  amount: number;
  beforePoints: number;
  afterPoints?: number | null;
  jobId?: string | null;
  data?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "lora-categories".
 */
export interface LoraCategory {
  id: string;
  name: string;
  description?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "lora-models".
 */
export interface LoraModel {
  canImageToImage: boolean;
  id: string;
  title: string;
  value: string;
  loraCategories?: (string | null) | LoraCategory;
  cover?: string | Media | null;
  description?: string | null;
  reference?: (string | PublishedImage)[] | null;
  basicSettings?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  trainingParams?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  workflow?: (string | null) | Workflow;
  authors?: (string | User)[] | null;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface PublishedImage {
  id: string;
  pointTransferLogs: string | PointTransferLog;
  user: string | User;
  url: string;
  is_published?: boolean | null;
  shortId?: string;
  reactions?:
    | {
        type: "like" | "dislike";
        count: number;
        id?: string | null;
      }[]
    | null;
  userReaction: { like: boolean; dislike: boolean; comment: string };
  updatedAt: string;
  createdAt: string;
}